Setup local environment with default stuff
After using commands that must be run OUTSIDE of a bash script, this bash script can be run on it's own to set up ideal/required tools for developers. Though future tools may (or, honestly - WILL BE - needed), this is a good place to start. Future versions of this file will surely include other important tools. For example, are wget, sudo, vim, curl, etc. already installed after this file or the files mentioned OUTSIDE of a bash script have run? No idea. But this is where commands for installing them will go until a better deployment solution for handling (either high-level or level) dependencies is found.